feat: add reduce count functionality to the order section
This commit adds the functionality to reduce the quantity of a product in an existing order. The `reduce_count` function in the `order.py` file now retrieves the order using the `OrderManager.from_id` method and reduces the count of the specified product using the `order.reduce_count` method. The `reduce_count` method is implemented in the `OrderManager` class in the `OrderManager.py` file, which calls the `Order.reduce_count_product` method in the `database.py` file to update the count of the order product.